requirements.txt
Django==3.2.8
psycopg2-binary==2.9.1
python-decouple==3.5
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-rq==2.4.1
rq==1.10.0
redis==3.5.3